# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FrontRow.AI — Musixmatch lyrics API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface Track {
  track_id: number;
  track_name: string;
  artist_name: string;
  album_name: string;
  /**
     * Spotify track ID when provided by the upstream payload; null when unavailable
     * @nullable
     */
  spotify_track_id: string | null;
  /**
     * Primary Musixmatch genre tag for the track; null when unavailable
     * @nullable
     */
  genre: string | null;
  /** @nullable */
  album_coverart_100x100?: string | null;
  /** @nullable */
  album_coverart_350x350?: string | null;
  /** 1 if word-level sync is available, 0 otherwise */
  has_richsync: number;
  /** 1 if line-level sync is available, 0 otherwise */
  has_subtitles: number;
  track_rating?: number;
  num_favourite?: number;
}

export interface RichsyncWord {
  /** Word/syllable text */
  c: string;
  /** Offset in seconds from line start */
  o: number;
}

export interface RichsyncLine {
  /** Line start time in seconds */
  ts: number;
  /** Line end time in seconds */
  te: number;
  /** Array of word-level entries */
  l: RichsyncWord[];
  /** Full line text */
  x: string;
}

export interface RichsyncResponse {
  track_id: number;
  /** Total lyrics duration in seconds */
  duration: number;
  lines: RichsyncLine[];
}

export interface SyncedLyricsLine {
  /** Line start time in seconds */
  time: number;
  /** Full line text */
  text: string;
}

export interface SyncedLyricsResponse {
  track_id: number;
  lines: SyncedLyricsLine[];
}

export interface PlainLyricsResponse {
  track_id: number;
  /** Full plain lyrics body, newline-separated lines (unsynced) */
  lyrics: string;
}

export type VenueType = typeof VenueType[keyof typeof VenueType];


export const VenueType = {
  indoor: 'indoor',
  outdoor: 'outdoor',
  amphitheater: 'amphitheater',
  symphony: 'symphony',
  festival: 'festival',
} as const;

export interface Show {
  id: string;
  /** Event or show name */
  name: string;
  venue_name: string;
  venue_type: VenueType;
  city: string;
  state: string;
  /** ISO date string from JamBase */
  date: string;
  /** @nullable */
  url?: string | null;
}

export interface ShowsResponse {
  artist: string;
  shows: Show[];
}

export type SearchTracksParams = {
/**
 * Search query (track title, artist name, or both)
 */
q: string;
/**
 * Page number
 */
page?: number;
/**
 * Results per page (max 100)
 */
page_size?: number;
};

export type GetShowsParams = {
/**
 * Artist name to search upcoming shows for
 */
artist: string;
};