# lib/api-zod/src/generated/types/plainLyricsResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FrontRow.AI — Musixmatch lyrics API
 * OpenAPI spec version: 0.1.0
 */

export interface PlainLyricsResponse {
  track_id: number;
  /** Full plain lyrics body, newline-separated lines (unsynced) */
  lyrics: string;
}