# lib/api-zod/src/generated/types/richsyncLine.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FrontRow.AI — Musixmatch lyrics API
 * OpenAPI spec version: 0.1.0
 */
import type { RichsyncWord } from './richsyncWord';

export interface RichsyncLine {
  /** Line start time in seconds */
  ts: number;
  /** Line end time in seconds */
  te: number;
  /** Array of word-level entries */
  l: RichsyncWord[];
  /** Full line text */
  x: string;
}