# lib/api-zod/src/generated/types/richsyncResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FrontRow.AI — Musixmatch lyrics API
 * OpenAPI spec version: 0.1.0
 */
import type { RichsyncLine } from './richsyncLine';

export interface RichsyncResponse {
  track_id: number;
  /** Total lyrics duration in seconds */
  duration: number;
  lines: RichsyncLine[];
}