# lib/api-zod/src/generated/types/richsyncWord.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FrontRow.AI — Musixmatch lyrics API
 * OpenAPI spec version: 0.1.0
 */

export interface RichsyncWord {
  /** Word/syllable text */
  c: string;
  /** Offset in seconds from line start */
  o: number;
}