# lib/api-zod/src/generated/types/show.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FrontRow.AI — Musixmatch lyrics API
 * OpenAPI spec version: 0.1.0
 */
import type { VenueType } from './venueType';

export interface Show {
  id: string;
  /** Event or show name */
  name: string;
  venue_name: string;
  venue_type: VenueType;
  city: string;
  state: string;
  /** ISO date string from JamBase */
  date: string;
  /** @nullable */
  url?: string | null;
}