# lib/api-zod/src/generated/types/showsResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FrontRow.AI — Musixmatch lyrics API
 * OpenAPI spec version: 0.1.0
 */
import type { Show } from './show';

export interface ShowsResponse {
  artist: string;
  shows: Show[];
}