# lib/api-zod/src/generated/types/syncedLyricsLine.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FrontRow.AI — Musixmatch lyrics API
 * OpenAPI spec version: 0.1.0
 */

export interface SyncedLyricsLine {
  /** Line start time in seconds */
  time: number;
  /** Full line text */
  text: string;
}