# lib/api-zod/src/generated/types/syncedLyricsResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FrontRow.AI — Musixmatch lyrics API
 * OpenAPI spec version: 0.1.0
 */
import type { SyncedLyricsLine } from './syncedLyricsLine';

export interface SyncedLyricsResponse {
  track_id: number;
  lines: SyncedLyricsLine[];
}