# lib/api-zod/src/generated/types/track.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FrontRow.AI — Musixmatch lyrics API
 * OpenAPI spec version: 0.1.0
 */

export interface Track {
  track_id: number;
  track_name: string;
  artist_name: string;
  album_name: string;
  /**
     * Spotify track ID when provided by the upstream payload; null when unavailable
     * @nullable
     */
  spotify_track_id: string | null;
  /**
     * Primary Musixmatch genre tag for the track; null when unavailable
     * @nullable
     */
  genre: string | null;
  /** @nullable */
  album_coverart_100x100?: string | null;
  /** @nullable */
  album_coverart_350x350?: string | null;
  /** 1 if word-level sync is available, 0 otherwise */
  has_richsync: number;
  /** 1 if line-level sync is available, 0 otherwise */
  has_subtitles: number;
  track_rating?: number;
  num_favourite?: number;
}