# lib/api-zod/src/generated/types/venueType.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FrontRow.AI — Musixmatch lyrics API
 * OpenAPI spec version: 0.1.0
 */

export type VenueType = typeof VenueType[keyof typeof VenueType];


export const VenueType = {
  indoor: 'indoor',
  outdoor: 'outdoor',
  amphitheater: 'amphitheater',
  symphony: 'symphony',
  festival: 'festival',
} as const;